Fixed the server status php
fixed the index file so that the server status php worked like it
should.  This meant takeing the .html and chaning it to .php to support
the use of php.

I also removed the old php files that were being used and moved all the
cotent to another file named status.php.

I also edited some of the content on the index page to reflect the
changes discussed in issue #6.
